Fixed bug with fields

+++ implementation/org.gravity.security.annotations.ui/src/main/java/org/gravity/security/annotations/marker/SecurityMarkerUtil.java
@@ -10,34 +10,34 @@ import org.eclipse.jdt.core.IMember;
 import org.gravity.eclipse.util.JavaASTUtil;
 
 public class SecurityMarkerUtil {
-	
+
 	public static final String MARKER_SOURCE = "org.gravity.security";
 	public static final String MARKER_ATTR_ANALYZED = "org.gravity.security.analyzedMember";
-	
+
 	/**
 	 * The logger of this class
 	 */
 	private static final Logger LOGGER = Logger.getLogger(SecurityMarkerUtil.class);
-	
+
 	private SecurityMarkerUtil() {
 		// This class should not be instantiated
 	}
 
-
-	public static IMarker createErrorMarker(IMember member, String message, String... relevantMember) {
+	public static IMarker createErrorMarker(final IMember member, final String message,
+			final String... relevantMember) {
 		try {
-			IResource resource = member.getResource();
-			int line = JavaASTUtil.getLine(member);
+			final var resource = member.getResource();
+			final var line = JavaASTUtil.getLine(member);
 
 			return createErrorMarker(resource, line, message, relevantMember);
-		} catch (CoreException e) {
+		} catch (final CoreException e) {
 			throw new IllegalStateException(e);
 		}
 	}
 
-	public static IMarker createErrorMarker(IResource resource, int line, String message, String... relevantMember)
-			throws CoreException {
-		IMarker m = resource.createMarker(IMarker.PROBLEM);
+	public static IMarker createErrorMarker(final IResource resource, final int line, final String message,
+			final String... relevantMember) throws CoreException {
+		final var m = resource.createMarker(IMarker.PROBLEM);
 		m.setAttribute(IMarker.LINE_NUMBER, line);
 		m.setAttribute(IMarker.MESSAGE, message);
 		m.setAttribute(IMarker.PRIORITY, IMarker.PRIORITY_HIGH);
@@ -47,47 +47,47 @@ public class SecurityMarkerUtil {
 		return m;
 	}
 
-	public static void createWarningMarker(IAnnotation annotation, String message) {
+	public static void createWarningMarker(final IAnnotation annotation, final String message) {
 		try {
-			IResource resource = annotation.getResource();
-			int line = JavaASTUtil.getLine(annotation);
+			final var resource = annotation.getResource();
+			final var line = JavaASTUtil.getLine(annotation);
 
-			IMarker m = resource.createMarker(IMarker.PROBLEM);
+			final var m = resource.createMarker(IMarker.PROBLEM);
 			m.setAttribute(IMarker.LINE_NUMBER, line);
 			m.setAttribute(IMarker.MESSAGE, message);
 			m.setAttribute(IMarker.PRIORITY, IMarker.PRIORITY_NORMAL);
 			m.setAttribute(IMarker.SEVERITY, IMarker.SEVERITY_WARNING);
 			m.setAttribute(IMarker.SOURCE_ID, MARKER_SOURCE);
-		} catch (CoreException e) {
+		} catch (final CoreException e) {
 			throw new IllegalStateException(e);
 		}
 	}
-	
-	public static void deleteOldMarkers(final ICompilationUnit cu) {
+
+	public static void deleteOldMarkers(final ICompilationUnit cu, final long timestamp) {
 		try {
-			for (IMarker m : cu.getResource().findMarkers(IMarker.PROBLEM, true, IResource.DEPTH_INFINITE)) {
-				if (MARKER_SOURCE.equals(m.getAttribute(IMarker.SOURCE_ID))) {
+			for (final IMarker m : cu.getResource().findMarkers(IMarker.PROBLEM, true, IResource.DEPTH_INFINITE)) {
+				if (m.getCreationTime() < timestamp && MARKER_SOURCE.equals(m.getAttribute(IMarker.SOURCE_ID))) {
 					m.delete();
 				}
 			}
-		} catch (CoreException e) {
+		} catch (final CoreException e) {
 			LOGGER.error(e);
 		}
 	}
 
-	public static void deleteOldMarkers(final IResource res, String analyzedMember) {
-		if (res != null)
+	public static void deleteOldMarkers(final IResource res, final String analyzedMember, final long timestamp) {
+		if (res != null) {
 			try {
-				for (IMarker m : res.findMarkers(IMarker.PROBLEM, true, IResource.DEPTH_INFINITE)) {
-					if (MARKER_SOURCE.equals(m.getAttribute(IMarker.SOURCE_ID))
+				for (final IMarker m : res.findMarkers(IMarker.PROBLEM, true, IResource.DEPTH_INFINITE)) {
+					if (m.getCreationTime() < timestamp && MARKER_SOURCE.equals(m.getAttribute(IMarker.SOURCE_ID))
 							&& m.getAttribute(MARKER_ATTR_ANALYZED, "").contains(analyzedMember)) {
 						m.delete();
 					}
 				}
-			} catch (CoreException e) {
+			} catch (final CoreException e) {
 				LOGGER.error(e);
 			}
+		}
 	}
 
-	
 }
